Can now ping withtout the localhost but on my computer

+++ includes/client_udp.hpp
@@ -32,7 +32,7 @@ class client_udp : public QObject
         QUdpSocket *udpsocket;
         QTimer timer;
         QHostAddress address;
-        int port;
+        quint16 port;
 };
 
 #endif //CLIENT_UDP
